Add rewrite test cases.
